Finished initial balancing operations

+++ headers/HeuristicStrategy.h
@@ -9,7 +9,7 @@ class Heuristic {
 
     public:
         Heuristic();
-        virtual int heuristic(const ShipState*) const = 0;
+        virtual double heuristic(const ShipState*) const = 0;
 };
 
 // Balance heuristics
@@ -17,7 +17,7 @@ class BalanceFactorHeuristic : public Heuristic {
 
     public:
         BalanceFactorHeuristic();
-        virtual int heuristic(const ShipState*) const override;
+        virtual double heuristic(const ShipState*) const override;
 };
 
 // Uniform cost heuristic
@@ -25,7 +25,7 @@ class UniformCostHeuristic : public Heuristic {
 
     public:
         UniformCostHeuristic();
-        virtual int heuristic(const ShipState*) const override;
+        virtual double heuristic(const ShipState*) const override;
 };
 
 #endif // HEURISTICSTRATEGY.H
